+++ libs/core-rest-api/adapters/src/controllers/api/use-cases/psychologist/authenticate-psychologist/authenticate-psychologist.e2e-spec.ts
@@ -1,72 +1,56 @@
 import request from 'supertest';
 
 import { INestApplication } from '@nestjs/common';
-import { Test } from '@nestjs/testing';
 
 import { fakerPT_BR as faker } from '@faker-js/faker';
-import { hash } from 'bcryptjs';
 
-import { PsychologistFactory } from '../../../../../../tests/factories/make-psychologist';
-import { DatabaseRepositoriesModule } from '../../../../../database/repositories/repositories.module';
-import { ApiModule } from '../../../api.module';
+import { PsychologistEntity } from '@clinicControl/core-rest-api/core/src/domains/psychologist/entities/psychologist/entity';
+import { setupE2ETest } from '../../../shared/utils/e2e-tests-initial-setup';
 
 describe('[E2E] -  Authenticate Psychologist', () => {
   let app: INestApplication;
-  let psychologistFactory: PsychologistFactory;
+
+  let psychologist: PsychologistEntity;
+  let password: string;
 
   beforeAll(async () => {
-    const moduleRef = await Test.createTestingModule({
-      imports: [ApiModule, DatabaseRepositoriesModule],
-      providers: [PsychologistFactory],
-    }).compile();
+    const setup = await setupE2ETest();
 
-    app = moduleRef.createNestApplication();
+    app = setup.app;
 
-    psychologistFactory = moduleRef.get(PsychologistFactory);
-
-    await app.init();
+    psychologist = setup.psychologist;
+    password = setup.password;
   });
 
   it('[POST] - Should successfully authenticate a psychologist', async () => {
-    const password = faker.internet.password({ length: 8 });
-    const newPsychologist = await psychologistFactory.makePrismaPsychologist({
-      password: await hash(password, 8),
-    });
-
     const response = await request(app.getHttpServer()).post('/psychologist/login').send({
-      email: newPsychologist.email,
+      email: psychologist.email,
       password,
     });
 
     expect(response.statusCode).toBe(201);
     expect(response.body).toEqual({
       user: {
-        id: newPsychologist.id,
-        name: newPsychologist.name,
-        email: newPsychologist.email,
+        id: psychologist.id,
+        name: psychologist.name,
+        email: psychologist.email,
       },
       access_token: expect.any(String),
     });
   });
 
   it('[POST] - Should return an error 401 when trying to authenticate with wrong password', async () => {
-    const password = faker.internet.password({ length: 8 });
-    const newPsychologist = await psychologistFactory.makePrismaPsychologist({
-      password: await hash(password, 8),
-      email: 'new_user_entrie@email.com',
-    });
-
     const response = await request(app.getHttpServer())
       .post('/psychologist/login')
       .send({
-        email: newPsychologist.email,
+        email: psychologist.email,
         password: faker.internet.password({ length: 3 }),
       });
 
     expect(response.statusCode).toBe(401);
   });
 
-  it('[POST] - Should return an error 401 when trying to authenticate with wrong password', async () => {
+  it('[POST] - Should return an error 401 when trying to authenticate with wrong email', async () => {
     const password = faker.internet.password({ length: 8 });
 
     const response = await request(app.getHttpServer()).post('/psychologist/login').send({
@@ -75,5 +59,6 @@ describe('[E2E] -  Authenticate Psychologist', () => {
     });
 
     expect(response.statusCode).toBe(401);
+    expect(response.body.message).toBe('psychologist not found');
   });
 });