+++ libs/core-rest-api/core/src/domains/clinic/repositories/database-in-memory-repository.ts
@@ -2,6 +2,7 @@ import { ConflictException } from '@nestjs/common';
 import { CLINIC_ERROR_MESSAGES } from '../../../shared/errors/error-messages';
 import { ClinicEntity } from '../../clinic/entities/clinic/entity';
 import { CreateClinicDto } from '../../clinic/use-cases/create-clinic/create-clinic-dto';
+import { DeletedClinicInfo } from '../use-cases/delete-clinic/dto';
 import { UpdateClinicDto } from '../use-cases/update-clinic/update-clinic-dto';
 import { ClinicDatabaseRepository } from './database-repository';
 
@@ -72,17 +73,19 @@ export class InMemoryClinicDatabaseRepository implements ClinicDatabaseRepositor
     this.clinics[clinicIndex] = updatedClinic;
   }
 
-  async deleteClinic(name: string, psychologistId: string): Promise<void> {
-    const isClinicExists = await this.findClinicByNameAndPsychologistId(
-      name,
-      psychologistId
+  async deleteClinic(id: string): Promise<DeletedClinicInfo> {
+    const isClinicExists = await this.findClinicById(
+      id
     );
 
     if (!isClinicExists) {
       throw new ConflictException(CLINIC_ERROR_MESSAGES['CLINIC_NOT_FOUND']);
     }
 
-    this.clinics = this.clinics.filter((clinic) => clinic.name !== name);
+    this.clinics = this.clinics.filter((clinic) => clinic.id !== id);
+    return {
+      deletedClinic: isClinicExists
+    }
   }
 
   async deleteAllClinicsByPsychologistId(