+++ libs/core-rest-api/core/src/domains/clinic/use-cases/delete-clinic/dto.ts
@@ -1,0 +1,5 @@
+import { ClinicEntity } from "../../entities/clinic/entity";
+
+export interface DeletedClinicInfo {
+  deletedClinic: ClinicEntity;
+}